Extract filtering functions into a testable utility.
