+++ app/dashboard/page.tsx
@@ -13,6 +13,7 @@ import CoffeeChatsTab from './components/CoffeeChatsTab';
 import EventsTab from './components/EventsTab';
 import OpenSourceTab from './components/OpenSourceTab';
 import { getApiHeaders } from '@/app/lib/api-helpers';
+import { getFilteredOpenSourceColumns } from './lib/open-source-filter';
 import type {
   Application,
   ApplicationStatus,
@@ -40,46 +41,6 @@ import {
   APPLICATION_COMPLETION_COLUMNS,
   EVENT_COMPLETION_COLUMNS,
 } from './components/types';
-import partnershipsData from '@/partnerships/partnerships.json';
-
-function normalizePartnerName(s: string | null | undefined): string {
-  return (s ?? '').trim().toLowerCase();
-}
-
-function buildPartnershipNameMatchSet(
-  partnershipId: number | null,
-  displayName: string | null,
-  availablePartnerships: Array<{ id: number; name: string }>,
-  fullPartnerships: Array<{ id: number; name: string }>
-): Set<string> {
-  const out = new Set<string>();
-  if (displayName) out.add(normalizePartnerName(displayName));
-  if (partnershipId != null) {
-    const fromJson = partnershipsData.partnerships.find(p => p.id === partnershipId)?.name;
-    if (fromJson) out.add(normalizePartnerName(fromJson));
-    const fromAvail = availablePartnerships.find(p => p.id === partnershipId)?.name;
-    if (fromAvail) out.add(normalizePartnerName(fromAvail));
-    const fromFull = fullPartnerships.find(p => p.id === partnershipId)?.name;
-    if (fromFull) out.add(normalizePartnerName(fromFull));
-  }
-  return out;
-}
-
-function openSourceDateForMonthFilter(entry: OpenSourceEntry): string | null {
-  return entry.dateModified ?? entry.dateCreated ?? null;
-}
-
-function parseIsoDate(value?: string | null): Date | null {
-  if (!value) return null;
-  const d = new Date(value);
-  if (Number.isNaN(d.getTime())) return null;
-  return d;
-}
-
-function entryDateForCompletedWindow(entry: OpenSourceEntry): Date | null {
-  return parseIsoDate(entry.dateCreated ?? entry.dateModified ?? null);
-}
-
 // ===== MOCK DATA FEATURE TOGGLE START =====
 // Toggle this flag or comment out the seeding effect below to disable mock data.
 const ENABLE_DASHBOARD_MOCKS = false;
@@ -1603,160 +1564,17 @@ const hasSeededMockDataRef = useRef(false);
   }, [eventColumns, eventsFilter, isWithinCurrentMonth]);
 
   const filteredOpenSourceColumns = useMemo(() => {
-    let filtered: Record<OpenSourceColumnId, OpenSourceEntry[]> = {
-      plan: [],
-      babyStep: [],
-      inProgress: [],
-      done: [],
-    };
-
-    // Browsing completed history should not be hidden by "This Month".
-    const effectiveTimeFilter: BoardTimeFilter = viewingCompletedPartnershipName ? 'allTime' : openSourceFilter;
-
-    if (effectiveTimeFilter === 'allTime') {
-      filtered = { ...openSourceColumns };
-    } else {
-      (Object.keys(openSourceColumns) as OpenSourceColumnId[]).forEach(columnId => {
-        if (effectiveTimeFilter === 'modifiedThisMonth') {
-          filtered[columnId] = openSourceColumns[columnId].filter(entry =>
-            isWithinCurrentMonth(openSourceDateForMonthFilter(entry))
-          );
-        }
-      });
-    }
-
-    const completedForView = viewingCompletedPartnershipName
-      ? completedPartnerships.find(
-          c => normalizePartnerName(c.partnershipName) === normalizePartnerName(viewingCompletedPartnershipName)
-        )
-      : undefined;
-
-    let nameSet: Set<string> | null = null;
-    if (viewingCompletedPartnershipName) {
-      nameSet = buildPartnershipNameMatchSet(
-        completedForView?.partnershipId ?? null,
-        viewingCompletedPartnershipName,
-        availablePartnerships,
-        fullPartnerships
-      );
-    } else if (selectedPartnership) {
-      nameSet = buildPartnershipNameMatchSet(
-        selectedPartnershipId,
-        selectedPartnership,
-        availablePartnerships,
-        fullPartnerships
-      );
-    }
-
-    if (nameSet && nameSet.size > 0) {
-      const isActiveView = Boolean(selectedPartnership && !viewingCompletedPartnershipName);
-      (Object.keys(filtered) as OpenSourceColumnId[]).forEach(columnId => {
-        filtered[columnId] = filtered[columnId].filter(entry => {
-          const n = normalizePartnerName(entry.partnershipName);
-          if (n === '') return isActiveView;
-          return nameSet!.has(n);
-        });
-      });
-    }
-
-    if (selectedPartnership && !viewingCompletedPartnershipName) {
-      const total = (Object.keys(filtered) as OpenSourceColumnId[]).reduce(
-        (acc, k) => acc + filtered[k as OpenSourceColumnId].length,
-        0
-      );
-      const timeTotal = (Object.keys(openSourceColumns) as OpenSourceColumnId[]).reduce(
-        (acc, k) => {
-          const arr =
-            effectiveTimeFilter === 'allTime'
-              ? openSourceColumns[k]
-              : openSourceColumns[k].filter(entry => isWithinCurrentMonth(openSourceDateForMonthFilter(entry)));
-          return acc + arr.length;
-        },
-        0
-      );
-      if (total === 0 && timeTotal > 0) {
-        const fallback: Record<OpenSourceColumnId, OpenSourceEntry[]> = {
-          plan: [],
-          babyStep: [],
-          inProgress: [],
-          done: [],
-        };
-        (Object.keys(openSourceColumns) as OpenSourceColumnId[]).forEach(columnId => {
-          fallback[columnId] =
-            effectiveTimeFilter === 'allTime'
-              ? [...openSourceColumns[columnId]]
-              : openSourceColumns[columnId].filter(entry =>
-                  isWithinCurrentMonth(openSourceDateForMonthFilter(entry))
-                );
-        });
-        return fallback;
-      }
-    }
-
-    if (viewingCompletedPartnershipName) {
-      const total = (Object.keys(filtered) as OpenSourceColumnId[]).reduce(
-        (acc, k) => acc + filtered[k as OpenSourceColumnId].length,
-        0
-      );
-      const start = parseIsoDate(completedForView?.startedAt ?? null);
-      const end = parseIsoDate(completedForView?.completedAt ?? null);
-      if (total === 0 && start) {
-        const windowFiltered: Record<OpenSourceColumnId, OpenSourceEntry[]> = {
-          plan: [],
-          babyStep: [],
-          inProgress: [],
-          done: [],
-        };
-        (Object.keys(openSourceColumns) as OpenSourceColumnId[]).forEach(columnId => {
-          windowFiltered[columnId] = openSourceColumns[columnId].filter(entry => {
-            const d = entryDateForCompletedWindow(entry);
-            if (!d) return false;
-            if (d < start) return false;
-            if (end && d > end) return false;
-            return true;
-          });
-        });
-        const windowTotal = (Object.keys(windowFiltered) as OpenSourceColumnId[]).reduce(
-          (acc, k) => acc + windowFiltered[k as OpenSourceColumnId].length,
-          0
-        );
-        if (windowTotal > 0) {
-          return windowFiltered;
-        }
-      }
-    }
-
-    if (!viewingCompletedPartnershipName) {
-      const finalTotal = (Object.keys(filtered) as OpenSourceColumnId[]).reduce(
-        (acc, k) => acc + filtered[k as OpenSourceColumnId].length,
-        0
-      );
-      if (finalTotal === 0) {
-        const base: Record<OpenSourceColumnId, OpenSourceEntry[]> = {
-          plan: [],
-          babyStep: [],
-          inProgress: [],
-          done: [],
-        };
-        (Object.keys(openSourceColumns) as OpenSourceColumnId[]).forEach(columnId => {
-          base[columnId] =
-            effectiveTimeFilter === 'allTime'
-              ? [...openSourceColumns[columnId]]
-              : openSourceColumns[columnId].filter(entry =>
-                  isWithinCurrentMonth(openSourceDateForMonthFilter(entry))
-                );
-        });
-        const baseTotal = (Object.keys(base) as OpenSourceColumnId[]).reduce(
-          (acc, k) => acc + base[k as OpenSourceColumnId].length,
-          0
-        );
-        if (baseTotal > 0) {
-          return base;
-        }
-      }
-    }
-
-    return filtered;
+    return getFilteredOpenSourceColumns({
+      openSourceColumns,
+      openSourceFilter,
+      selectedPartnership,
+      selectedPartnershipId,
+      viewingCompletedPartnershipName,
+      completedPartnerships,
+      availablePartnerships,
+      fullPartnerships,
+      isWithinCurrentMonth,
+    });
   }, [
     openSourceColumns,
     openSourceFilter,
